requirements.txt
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
blinker==1.4
certifi==2018.1.18
cffi==1.12.2
chardet==3.0.4
click==6.7
colorama==0.3.7
configobj==5.0.6
constantly==15.1.0
coverage==4.5.2
cryptography==2.6.1
cycler==0.10.0
Django==2.1.7
django-extensions==2.1.5
django-nose==1.4.6
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
incremental==16.10.1
Jinja2==2.10
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.0.1
MarkupSafe==1.0
matplotlib==3.0.2
mpld3==0.3
nose==1.3.7
numpy==1.16.1
oauthlib==2.0.6
pandas==0.24.1
Pillow==5.4.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.19
pycrypto==2.6.1
PyJWT==1.5.3
pyOpenSSL==17.5.0
pyparsing==2.3.1
pyserial==3.4
python-dateutil==2.8.0
python-debian==0.1.32
pytz==2018.9
pyxdg==0.25
PyYAML==3.12
requests==2.18.4
requests-unixsocket==0.1.5
scikit-learn==0.20.2
scipy==1.2.1
SecretStorage==2.3.1
selenium==3.141.0
service-identity==16.0.0
six==1.12.0
sklearn==0.0
Twisted==17.9.0
urllib3==1.22
virtualenv==15.1.0
zope.interface==4.3.2
